+++ hub/spec/factories/data_products.rb
@@ -1,9 +1,10 @@
 FactoryBot.define do
   factory :data_product do
     target
-    url { "https://example-bucket.s3.amazonaws.com/sub_0001.fits" }
-    kind { :sub }
+    kind { :night_master }
+    sequence(:altair_id) { |n| n }
     filter { "Luminance" }
+    night { Date.current }
     captured_at { Time.current }
   end
 end